#33 Move UpdateParam outside insert
