+++ cql-ast/src/main/scala/troy/cql/ast/dml/UpdateParam.scala
@@ -1,0 +1,11 @@
+package troy.cql.ast.dml
+
+import troy.cql.ast.BindMarker
+
+sealed trait UpdateParam
+case class Timestamp(value: UpdateParamValue) extends UpdateParam
+case class Ttl(value: UpdateParamValue) extends UpdateParam
+
+sealed trait UpdateParamValue
+case class UpdateValue(value: String) extends UpdateParamValue
+case class UpdateVariable(bindMarker: BindMarker) extends UpdateParamValue
